Move the device test to test_init and tick by the scan interval
