+++ tests/components/sunsynk/snapshots/test_sensor.ambr
@@ -1,92 +1,4 @@
 # serializer version: 1
-# name: test_devices
-  list([
-    DeviceRegistryEntrySnapshot({
-      'area_id': None,
-      'config_entry_id': <ANY>,
-      'config_subentry_id': <ANY>,
-      'configuration_url': None,
-      'connections': set({
-      }),
-      'disabled_by': None,
-      'entry_type': None,
-      'hw_version': None,
-      'id': <ANY>,
-      'identifiers': set({
-        tuple(
-          'sunsynk',
-          '1029384756',
-        ),
-      }),
-      'labels': set({
-      }),
-      'manufacturer': 'Sunsynk',
-      'model': 'SUNSYNK-5K-SG04LP1',
-      'model_id': None,
-      'name': 'Garage inverter',
-      'name_by_user': None,
-      'serial_number': '1029384756',
-      'sw_version': '1.5.1.5',
-      'via_device_id': None,
-    }),
-    DeviceRegistryEntrySnapshot({
-      'area_id': None,
-      'config_entry_id': <ANY>,
-      'config_subentry_id': <ANY>,
-      'configuration_url': None,
-      'connections': set({
-      }),
-      'disabled_by': None,
-      'entry_type': None,
-      'hw_version': None,
-      'id': <ANY>,
-      'identifiers': set({
-        tuple(
-          'sunsynk',
-          '2938475610',
-        ),
-      }),
-      'labels': set({
-      }),
-      'manufacturer': 'Sunsynk',
-      'model': None,
-      'model_id': None,
-      'name': '2938475610',
-      'name_by_user': None,
-      'serial_number': '2938475610',
-      'sw_version': None,
-      'via_device_id': None,
-    }),
-    DeviceRegistryEntrySnapshot({
-      'area_id': None,
-      'config_entry_id': <ANY>,
-      'config_subentry_id': <ANY>,
-      'configuration_url': None,
-      'connections': set({
-      }),
-      'disabled_by': None,
-      'entry_type': None,
-      'hw_version': None,
-      'id': <ANY>,
-      'identifiers': set({
-        tuple(
-          'sunsynk',
-          '1029384756_battery',
-        ),
-      }),
-      'labels': set({
-      }),
-      'manufacturer': 'Sunsynk',
-      'model': None,
-      'model_id': None,
-      'name': 'Battery',
-      'name_by_user': None,
-      'serial_number': None,
-      'sw_version': None,
-      'via_device_id': <ANY>,
-    }),
-  ])
-# ---
 # name: test_sensors[sensor.2938475610_grid_export_today-entry]
   EntityRegistryEntrySnapshot({
     'aliases': list([
